Add fdroid build flavor
- Move GMS depending code to gms flavor.
- Add Wrappers for functions depending on GMS.
- Add empty implementations for fdroid flavor.
  TODO: provide non GMS implementations.

Closes: #70
