+++ app/build.gradle
@@ -2,7 +2,6 @@ plugins {
     id 'com.android.application'
     id 'kotlin-android'
     id 'kotlin-kapt'
-    id 'com.google.gms.google-services'
     id 'androidx.navigation.safeargs.kotlin'
 }
 
@@ -52,6 +51,13 @@ android {
     kotlinOptions {
         jvmTarget = "1.8"
     }
+    flavorDimensions "default"
+    productFlavors {
+        create("gms") {
+        }
+        create("fdroid") {
+        }
+    }
 }
 
 dependencies {
@@ -80,13 +86,13 @@ dependencies {
     implementation 'com.github.AppIntro:AppIntro:6.1.0'
 
     // Fused Location Client
-    implementation 'com.google.android.gms:play-services-location:18.0.0'
+    gmsImplementation 'com.google.android.gms:play-services-location:18.0.0'
 
     // Firebase
     // Import the BoM for the Firebase platform
-    implementation platform('com.google.firebase:firebase-bom:26.4.0')
+    gmsImplementation platform('com.google.firebase:firebase-bom:26.4.0')
 
-    implementation 'com.google.firebase:firebase-messaging-ktx:21.0.1'
+    gmsImplementation 'com.google.firebase:firebase-messaging-ktx:21.0.1'
 
     // Settings
     implementation 'androidx.preference:preference-ktx:1.1.1'
@@ -124,3 +130,6 @@ dependencies {
     debugImplementation 'com.squareup.leakcanary:leakcanary-android:2.7'
 }
 
+if (getGradle().getStartParameter().getTaskRequests().toString().contains("Gms")) {
+    apply plugin: 'com.google.gms.google-services'
+}
