+++ app/src/fdroid/java/com/meteocool/firebase/MyFirebaseMessagingService.kt
@@ -1,0 +1,12 @@
+package com.meteocool.firebase
+
+import android.content.Context
+
+
+class MyFirebaseMessagingService {
+
+    companion object {
+        fun cancelNotification(context: Context, from : String) {
+        }
+    }
+}
