+++ app/src/fdroid/java/com/meteocool/location/ResolvableApiException.kt
@@ -1,0 +1,7 @@
+package com.meteocool.location
+
+import android.app.Activity
+
+class ResolvableApiException: Exception() {
+    fun startResolutionForResult(activity: Activity, requestCode: Int) {}
+}
