+++ app/src/fdroid/java/com/meteocool/location/service/LocationServiceFactory.kt
@@ -1,0 +1,14 @@
+package com.meteocool.location.service
+
+import android.content.Context
+
+
+class LocationServiceFactory {
+
+    companion object {
+
+        fun getLocationService(context : Context) : ForegroundLocationService {
+             return FusedLocationService(context)
+        }
+    }
+}
