+++ app/src/fdroid/java/com/meteocool/preferences/FirebaseMessagingWrapper.kt
@@ -1,0 +1,11 @@
+package com.meteocool.preferences
+
+class FirebaseMessagingWrapper {
+
+    companion object {
+
+        fun getFirebaseToken(): String {
+            return ""
+        }
+    }
+}
