+++ app/src/main/java/com/meteocool/preferences/SettingsFragment.kt
@@ -14,8 +14,7 @@ import androidx.lifecycle.Observer
 import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
 import androidx.preference.SwitchPreferenceCompat
-import com.google.android.gms.tasks.OnCompleteListener
-import com.google.firebase.messaging.FirebaseMessaging
+import com.meteocool.preferences.FirebaseMessagingWrapper
 import com.meteocool.R
 import com.meteocool.injection.InjectorUtils
 import com.meteocool.network.NetworkUtils
@@ -102,30 +101,27 @@ class SettingsFragment() : PreferenceFragmentCompat() {
     private fun registerPreferenceClickListener() {
         findPreference<Preference>("feedback")?.setOnPreferenceClickListener {
             val tokenInShared = SharedPrefUtils.getFirebaseToken(defaultSharedPreferences)
-            FirebaseMessaging.getInstance().token.addOnCompleteListener(OnCompleteListener { task ->
-                if (!task.isSuccessful) {
-                    handleExternalLink(
-                        getString(R.string.feedback_url) + "\n" + "Token fetch failed\nShared-Token: $tokenInShared"
-                    )
-                    return@OnCompleteListener
-                }
-
-                val token = task.result
-                var version: String = ""
-                try {
-                    val pInfo = requireContext().packageManager.getPackageInfo(
-                        requireContext().packageName,
-                        0
-                    )
-                    version = pInfo.versionName
-                } catch (e: PackageManager.NameNotFoundException) {
-                    e.printStackTrace()
-                }
+            val token = FirebaseMessagingWrapper.getFirebaseToken()
+            if (token == "no token") {
                 handleExternalLink(
-                    getString(R.string.feedback_url) + "\n" + token + "\nShared-Token: $tokenInShared" + "\n" + version
-
+                    getString(R.string.feedback_url) + "\n" + "Token fetch failed\nShared-Token: $tokenInShared"
                 )
-            })
+            }
+
+            var version: String = ""
+            try {
+                val pInfo = requireContext().packageManager.getPackageInfo(
+                    requireContext().packageName,
+                    0
+                )
+                version = pInfo.versionName
+            } catch (e: PackageManager.NameNotFoundException) {
+                e.printStackTrace()
+            }
+            handleExternalLink(
+                getString(R.string.feedback_url) + "\n" + token + "\nShared-Token: $tokenInShared" + "\n" + version
+
+            )
             true
         }
         findPreference<Preference>("github")?.setOnPreferenceClickListener {
